event listeners added to index, runner logic separated out
