+++ lib/runner.js
@@ -3,6 +3,8 @@ var $ = require('jquery');
 
 module.exports = function Runner() {
     this.ideasList = [];
+    this.uncompletedIdeasList = [];
+
 
     Runner.prototype.generateNewIdea = function (titleInput, bodyInput) {
       var idea = new Idea(titleInput, bodyInput);
@@ -33,23 +35,40 @@ module.exports = function Runner() {
           return ideasToKeep != idea;
         });
         localStorage.removeItem(idea);
-        updateIdeasList(this.ideasList);
+        this.updateIdeasList(this.ideasList);
     };
 
     Runner.prototype.updateIdeasList = function (ideasList) {
         localStorage.setItem('ideasList', JSON.stringify(this.ideasList));
-        storeIdea();
+        this.storeIdea();
     };
 
   Runner.prototype.updateBody = function (id, newBody) {
       var idea = findIdea(id);
       this.idea.body = newBody;
-      storeIdea();
+      this.storeIdea();
     };
-Runner.prototype.updateTitle = function (id, newTitle) {
-  var idea = findIdea(id);
-  this.idea.title = newTitle;
-  storeIdea();
-};
 
+  Runner.prototype.updateTitle = function (id, newTitle) {
+    var idea = findIdea(id);
+    this.idea.title = newTitle;
+    this.storeIdea();
+  };
+
+  Runner.prototype.writeIdeas = function (ideasList) {
+    this.ideasList.forEach(function(idea) {
+      if (this.idea.completed === false) {
+      this.idea.renderUncompletedIdeaToHTML(idea);
+      }
+    });
+
+    Runner.prototype.numIdeas = function() {
+      var numIdeas = this.ideasList.length;
+      var begin = 0;
+
+      slicedIdeas = this. ideasList.slice(begin, end);
+      this.writeIdeas(slicedIdeas);
+    };
+
+  };
 };
