+++ lib/scripts.js
@@ -1,5 +1,5 @@
 // var ideasList = [];
-var uncompletedIdeasList = [];
+// var uncompletedIdeasList = [];
 // var $ = require('jQuery');
 var end = 10;
 
@@ -99,150 +99,150 @@ var end = 10;
 //   storeIdea();
 // }
 
-function writeIdeas(ideasList) {
-  ideasList.forEach(function(idea) {
-    if (idea.completed === false) {
-    renderUncompletedIdeaToHTML(idea);
-    }
-  });
-}
+// (runner)function writeIdeas(ideasList) {
+//   ideasList.forEach(function(idea) {
+//     if (idea.completed === false) {
+//     renderUncompletedIdeaToHTML(idea);
+//     }
+//   });
+// }
 
-function numIdeas() {
-  var numIdeas = ideasList.length;
-  var begin = 0;
+// function numIdeas() {
+//   var numIdeas = ideasList.length;
+//   var begin = 0;
+//
+//   slicedIdeas = ideasList.slice(begin, end);
+//   writeIdeas(slicedIdeas);
+// }
+// $('#title-input').on('keypress', function() {
+//   if ($(this).val() === "" || $(this).val().length > 120){
+//     $('#btn-save').prop('disabled', true);
+//   }else {    $('#btn-save').prop('disabled', false);
+// }
+// });
+// $(document).ready(function() {
+//   ideasList = JSON.parse(localStorage.getItem('ideasList')) || [];
+//   // disable submit button
+//   $('#btn-save').prop('disabled', true);
+//   writeIdeas(ideasList);
+// });
 
-  slicedIdeas = ideasList.slice(begin, end);
-  writeIdeas(slicedIdeas);
-}
-$('#title-input').on('keypress', function() {
-  if ($(this).val() === "" || $(this).val().length > 120){
-    $('#btn-save').prop('disabled', true);
-  }else {    $('#btn-save').prop('disabled', false);
-}
-});
-$(document).ready(function() {
-  ideasList = JSON.parse(localStorage.getItem('ideasList')) || [];
-  // disable submit button
-  $('#btn-save').prop('disabled', true);
-  writeIdeas(ideasList);
-});
-
-$('#nextTen').on('click', function(){
-  $('.list-li').remove();
-  numIdeas();
-  end = end + 10;
-});
+// $('#nextTen').on('click', function(){
+//   $('.list-li').remove();
+//   numIdeas();
+//   end = end + 10;
+// });
 
 // (idea)function criticalIdeas(idea){
 //   return idea.quality === 'Critical';
 // }
 
-$('#btn-critical').on('click', function(){
-  $('.list-li').remove();
-  //when clicked, remove current todos
-  criticalIdeas =ideasList.filter(criticalIdeas)
-  writeIdeas(criticalIdeas);
+// $('#btn-critical').on('click', function(){
+//   $('.list-li').remove();
+//   //when clicked, remove current todos
+//   criticalIdeas =ideasList.filter(criticalIdeas)
+//   writeIdeas(criticalIdeas);
   //filter through todos
   //return critical ideas
   //write critical ideas to page
 });
 
 
-$('.idea-list').on('focusout', '.body-input', function(){
-  var id = $(this).parent().attr('id');
-  var newBody =  $(this).text();
-  updateBody(id, newBody);
-});
+// $('.idea-list').on('focusout', '.body-input', function(){
+//   var id = $(this).parent().attr('id');
+//   var newBody =  $(this).text();
+//   updateBody(id, newBody);
+// });
 
-$('.idea-list').on('keypress', '.body-input', function(event) {
-  var id = $(this).parent().attr('id');
-  var newBody = $(this).text();
-  if (event.which == 13) {
-    event.preventDefault();
-    $(this).blur();
-  }
-});
+// $('.idea-list').on('keypress', '.body-input', function(event) {
+//   var id = $(this).parent().attr('id');
+//   var newBody = $(this).text();
+//   if (event.which == 13) {
+//     event.preventDefault();
+//     $(this).blur();
+//   }
+// });
 
-$('.idea-list').on('focusout', '.idea-title', function(){
-  var id = $(this).parent().attr('id');
-  var newTitle =  $(this).text();
-  updateTitle(id, newTitle);
-});
-
-$('.idea-list').on('keypress', '.idea-title', function(event) {
-  var id = $(this).parent().attr('id');
-  var newTitle = $(this).text();
-  if (event.which == 13) {
-    event.preventDefault();
-    $(this).blur();
-  }
-});
-
-$('#title-input').on('keyup', function () {
-  var newTitle =  $(this).val();
-  //count characters of new title
-  charLength = newTitle.length++;
-
-  $('#char-title').text("Character counter: " + charLength);
-});
-
-$('.idea-list').on('click', '.upvote', function() {
-  var idea = findIdea($(this).parent().parent().attr('id'));
-  var $quality = $(this).siblings('p');
-
-  if ($quality.text() === 'quality: None') {
-     $quality.text('quality: Low');
-     idea.quality = 'Low';
-  } else if ($quality.text() === 'quality: Low') {
-     $quality.text('quality: Normal');
-     idea.quality = 'Normal';
-  } else if ($quality.text() === 'quality: Normal') {
-     $quality.text('quality: High');
-     idea.quality = 'High';
-  } else if ($quality.text() === 'quality: High') {
-     $quality.text('quality: Critical');
-     idea.quality = 'Critical';
-  }
-  storeIdea();
-});
-
-$('.idea-list').on('click', '.delete-idea', function(){
-  var id = $(this).parent().attr('id');
-  var idea =  findIdea(id);
-  deleteIdeaFromStorage(idea);
-  $(this).parent().remove();
-});
-
-$('.idea-list').on('click', '.completed-todo', function(){
-  var id = $(this).parent().parent().attr('id');
-  var idTag = "#" + id;
-  var idea =  findIdea(id);
-  markAsComplete(idea);
-  $('#'+id).children().addClass('completed');
-  storeIdea();
-});
-
-$('.idea-list').on('click', '.downvote', function() {
-  var idea = findIdea($(this).parent().parent().attr('id'));
-  var $quality = $(this).siblings('p');
-
-  if ($quality.text() === 'quality: Critical') {
-    $quality.text('quality: High');
-    idea.quality = 'High';
-  } else if ($quality.text() === 'quality: High') {
-    $quality.text('quality: Normal');
-    idea.quality = 'Normal';
-  } else if ($quality.text() === 'quality: Normal') {
-    $quality.text('quality: Low');
-    idea.quality = 'Low';
-  } else if ($quality.text() === 'quality: Low') {
-    $quality.text('quality: None');
-    idea.quality = 'None';
-  }
-  storeIdea();
-});
-
-//(idea) function completedIdeas(idea) {
+// $('.idea-list').on('focusout', '.idea-title', function(){
+//   var id = $(this).parent().attr('id');
+//   var newTitle =  $(this).text();
+//   updateTitle(id, newTitle);
+// });
+//
+// $('.idea-list').on('keypress', '.idea-title', function(event) {
+//   var id = $(this).parent().attr('id');
+//   var newTitle = $(this).text();
+//   if (event.which == 13) {
+//     event.preventDefault();
+//     $(this).blur();
+//   }
+// });
+//
+// $('#title-input').on('keyup', function () {
+//   var newTitle =  $(this).val();
+//   //count characters of new title
+//   charLength = newTitle.length++;
+//
+//   $('#char-title').text("Character counter: " + charLength);
+// });
+//
+// $('.idea-list').on('click', '.upvote', function() {
+//   var idea = findIdea($(this).parent().parent().attr('id'));
+//   var $quality = $(this).siblings('p');
+//
+//   if ($quality.text() === 'quality: None') {
+//      $quality.text('quality: Low');
+//      idea.quality = 'Low';
+//   } else if ($quality.text() === 'quality: Low') {
+//      $quality.text('quality: Normal');
+//      idea.quality = 'Normal';
+//   } else if ($quality.text() === 'quality: Normal') {
+//      $quality.text('quality: High');
+//      idea.quality = 'High';
+//   } else if ($quality.text() === 'quality: High') {
+//      $quality.text('quality: Critical');
+//      idea.quality = 'Critical';
+//   }
+//   storeIdea();
+// });
+//
+// $('.idea-list').on('click', '.delete-idea', function(){
+//   var id = $(this).parent().attr('id');
+//   var idea =  findIdea(id);
+//   deleteIdeaFromStorage(idea);
+//   $(this).parent().remove();
+// });
+//
+// $('.idea-list').on('click', '.completed-todo', function(){
+//   var id = $(this).parent().parent().attr('id');
+//   var idTag = "#" + id;
+//   var idea =  findIdea(id);
+//   markAsComplete(idea);
+//   $('#'+id).children().addClass('completed');
+//   storeIdea();
+// });
+//
+// $('.idea-list').on('click', '.downvote', function() {
+//   var idea = findIdea($(this).parent().parent().attr('id'));
+//   var $quality = $(this).siblings('p');
+//
+//   if ($quality.text() === 'quality: Critical') {
+//     $quality.text('quality: High');
+//     idea.quality = 'High';
+//   } else if ($quality.text() === 'quality: High') {
+//     $quality.text('quality: Normal');
+//     idea.quality = 'Normal';
+//   } else if ($quality.text() === 'quality: Normal') {
+//     $quality.text('quality: Low');
+//     idea.quality = 'Low';
+//   } else if ($quality.text() === 'quality: Low') {
+//     $quality.text('quality: None');
+//     idea.quality = 'None';
+//   }
+//   storeIdea();
+// });
+//
+// //(idea) function completedIdeas(idea) {
 //   return idea.completed === true;
 // }
 //
@@ -250,44 +250,44 @@ $('.idea-list').on('click', '.downvote', function() {
 //   return idea.completed === false;
 // }
 
-$('#btn-completed').on('click', function() {
-  ideasList = JSON.parse(localStorage.getItem('ideasList')) || [];
-  completedList = ideasList.filter(completedIdeas);
-
-  uncompletedList = ideasList.filter(uncompletedIdeas);
-  $('.list-li').remove();
-  uncompletedList.forEach(function(idea){
-    renderUncompletedIdeaToHTML(idea);
-  });
-  completedList.forEach(function(idea) {
-    renderCompletedIdeaToHTML(idea);
-  });
-
-});
-
-$( "#search-bar" ).keyup(function() {
-  var filterWord = $(this).val();
-  var notTheIdeasIWant = $( 'li:not(:contains(' + filterWord + '))' );
-  var theIdeaIWant = $('li:contains(' + filterWord + ')'
-  );
-  theIdeaIWant.show();
-  notTheIdeasIWant.hide();
-});
-
-$('#btn-save').on('click', function() {
-  var titleInput = $('#title-input').val();
-  var bodyInput = $('#body-input').val();
-  generateNewIdea(titleInput, bodyInput);
-  $('#title-input').focus();
-});
-
-$('#body-input').keypress(function(event) {
-  if (event.which == 13) {
-    var titleInput = $('#title-input').val();
-    var bodyInput = $('#body-input').val();
-    generateNewIdea(titleInput, bodyInput);
-    $('#title-input').focus();
-  }
-});
-
-module.exports = Idea;
+// $('#btn-completed').on('click', function() {
+//   ideasList = JSON.parse(localStorage.getItem('ideasList')) || [];
+//   completedList = ideasList.filter(completedIdeas);
+//
+//   uncompletedList = ideasList.filter(uncompletedIdeas);
+//   $('.list-li').remove();
+//   uncompletedList.forEach(function(idea){
+//     renderUncompletedIdeaToHTML(idea);
+//   });
+//   completedList.forEach(function(idea) {
+//     renderCompletedIdeaToHTML(idea);
+//   });
+//
+// });
+//
+// $( "#search-bar" ).keyup(function() {
+//   var filterWord = $(this).val();
+//   var notTheIdeasIWant = $( 'li:not(:contains(' + filterWord + '))' );
+//   var theIdeaIWant = $('li:contains(' + filterWord + ')'
+//   );
+//   theIdeaIWant.show();
+//   notTheIdeasIWant.hide();
+// });
+//
+// $('#btn-save').on('click', function() {
+//   var titleInput = $('#title-input').val();
+//   var bodyInput = $('#body-input').val();
+//   generateNewIdea(titleInput, bodyInput);
+//   $('#title-input').focus();
+// });
+//
+// $('#body-input').keypress(function(event) {
+//   if (event.which == 13) {
+//     var titleInput = $('#title-input').val();
+//     var bodyInput = $('#body-input').val();
+//     generateNewIdea(titleInput, bodyInput);
+//     $('#title-input').focus();
+//   }
+// });
+//
+// module.exports = Idea;
